requirements.txt
# Telegram Bot Framework
aiogram==3.4.1

# HTTP and Async
aiohttp==3.9.3
aiofiles==23.2.1
aiosignal==1.3.1

# Google AI (Gemini)
google-generativeai==0.8.5

# Database
asyncpg==0.29.0

# Environment Variables
python-dotenv==1.0.0

# Data Validation
pydantic==2.5.3
pydantic_core==2.14.6

# Utilities
attrs==23.2.0
certifi==2024.2.2
frozenlist==1.4.1
idna==3.6
magic-filter==1.0.12
multidict==6.0.5
typing_extensions==4.10.0
yarl==1.9.4
annotated-types==0.6.0

# Additional dependencies for better functionality
httpx==0.25.2
